+++ src/components/setting/developer/View.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
-import PluginSettings from '../plugins/View';
+import PluginSettings from './plugins/View';
 import { useNoxSetting } from '@stores/useApp';
 import SelectDialogWrapper from '../SelectDialogWrapper';
 import { Route } from './enums';
